Tutorials_Medium/1. Server/otel_config.py
# otel_config.py
from opentelemetry import trace
from opentelemetry.sdk.trace import TracerProvider
from opentelemetry.sdk.trace.export import (
    BatchSpanProcessor,
    ConsoleSpanExporter,
)

def setup_tracing():
    # Set up the tracer provider
    trace.set_tracer_provider(TracerProvider())

    # Create a span processor and exporter
    span_processor = BatchSpanProcessor(ConsoleSpanExporter())
    trace.get_tracer_provider().add_span_processor(span_processor)
